# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrainingGoal = (typeof TrainingGoal)[keyof typeof TrainingGoal];

export const TrainingGoal = {
  mountain_hiking: "mountain_hiking",
  heat_tolerance: "heat_tolerance",
  recovery: "recovery",
  speed_workout: "speed_workout",
  endurance: "endurance",
  general_fitness: "general_fitness",
} as const;

export type SurfaceType = (typeof SurfaceType)[keyof typeof SurfaceType];

export const SurfaceType = {
  pavement: "pavement",
  gravel: "gravel",
  trail: "trail",
  mixed: "mixed",
} as const;

export type RouteRequestTimeOfDay =
  (typeof RouteRequestTimeOfDay)[keyof typeof RouteRequestTimeOfDay];

export const RouteRequestTimeOfDay = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
  night: "night",
} as const;

export interface RouteRequest {
  trainingGoal: TrainingGoal;
  /**
   * @minimum 0.5
   * @maximum 30
   */
  distanceMiles: number;
  startLat: number;
  startLng: number;
  temperatureF?: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  humidity?: number;
  windSpeedMph?: number;
  /**
   * @minimum 0
   * @maximum 11
   */
  uvIndex?: number;
  timeOfDay?: RouteRequestTimeOfDay;
  preferShade?: boolean;
  avoidTraffic?: boolean;
  preferTrails?: boolean;
}

export interface ScoreBreakdown {
  terrainMatch: number;
  safetyScore: number;
  environmentalFit: number;
  trainingEffectiveness: number;
  shadeScore: number;
  trafficScore: number;
}

export interface Waypoint {
  lat: number;
  lng: number;
  elevation: number;
  name?: string;
}

export type RouteSegmentTrafficLevel =
  (typeof RouteSegmentTrafficLevel)[keyof typeof RouteSegmentTrafficLevel];

export const RouteSegmentTrafficLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface RouteSegment {
  startIndex: number;
  endIndex: number;
  distanceMiles: number;
  elevationGainFt: number;
  grade: number;
  surfaceType: SurfaceType;
  shadePercentage: number;
  trafficLevel: RouteSegmentTrafficLevel;
  /**
   * @minimum 0
   * @maximum 10
   */
  safetyScore: number;
  lightingAvailable: boolean;
}

export type GeneratedRouteSurfaceBreakdown = { [key: string]: number };

export interface GeneratedRoute {
  id: string;
  name: string;
  description: string;
  distanceMiles: number;
  estimatedDurationMinutes: number;
  elevationGainFt: number;
  elevationLossFt: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  overallScore: number;
  scoreBreakdown: ScoreBreakdown;
  waypoints: Waypoint[];
  segments: RouteSegment[];
  surfaceBreakdown: GeneratedRouteSurfaceBreakdown;
  warnings: string[];
  highlights: string[];
}

export interface WeatherSummary {
  temperatureF: number;
  humidity: number;
  windSpeedMph: number;
  uvIndex: number;
  conditions: string;
  heatIndex: number;
  recommendation: string;
}

export interface RouteResponse {
  routes: GeneratedRoute[];
  weatherSummary: WeatherSummary;
  generatedAt: string;
}

export type UserProfileHeatTolerance =
  (typeof UserProfileHeatTolerance)[keyof typeof UserProfileHeatTolerance];

export const UserProfileHeatTolerance = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export type UserProfileElevationTolerance =
  (typeof UserProfileElevationTolerance)[keyof typeof UserProfileElevationTolerance];

export const UserProfileElevationTolerance = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface UserProfile {
  id: number;
  nickname: string;
  preferredGoals: TrainingGoal[];
  preferredDistanceMiles: number;
  averagePaceMinsPerMile: number;
  preferredSurfaces: SurfaceType[];
  heatTolerance: UserProfileHeatTolerance;
  elevationTolerance: UserProfileElevationTolerance;
  totalRunsLogged: number;
  totalMilesRun: number;
  createdAt: string;
}

export type UserProfileUpdateHeatTolerance =
  (typeof UserProfileUpdateHeatTolerance)[keyof typeof UserProfileUpdateHeatTolerance];

export const UserProfileUpdateHeatTolerance = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export type UserProfileUpdateElevationTolerance =
  (typeof UserProfileUpdateElevationTolerance)[keyof typeof UserProfileUpdateElevationTolerance];

export const UserProfileUpdateElevationTolerance = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface UserProfileUpdate {
  nickname?: string;
  preferredGoals?: TrainingGoal[];
  preferredDistanceMiles?: number;
  averagePaceMinsPerMile?: number;
  preferredSurfaces?: SurfaceType[];
  heatTolerance?: UserProfileUpdateHeatTolerance;
  elevationTolerance?: UserProfileUpdateElevationTolerance;
}

export interface RunRecord {
  id: number;
  routeId: string;
  trainingGoal: TrainingGoal;
  distanceMiles: number;
  durationMinutes: number;
  avgPaceMinsPerMile: number;
  elevationGainFt: number;
  temperatureF: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  perceivedEffort: number;
  notes?: string;
  completedAt: string;
}

export interface LogRunRequest {
  routeId: string;
  trainingGoal: TrainingGoal;
  distanceMiles: number;
  durationMinutes: number;
  elevationGainFt: number;
  temperatureF: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  perceivedEffort: number;
  notes?: string;
}

export interface RunListResponse {
  runs: RunRecord[];
  total: number;
}

export type GoalWeightsWeights = { [key: string]: number };

export interface GoalWeights {
  label: string;
  description: string;
  weights: GoalWeightsWeights;
}

export type ScoringFactorsResponseGoals = { [key: string]: GoalWeights };

export interface ScoringFactorsResponse {
  goals: ScoringFactorsResponseGoals;
}

export type ListRunsParams = {
  limit?: number;
  offset?: number;
};