# lib/api-zod/src/generated/types/generatedRoute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedRouteSurfaceBreakdown } from "./generatedRouteSurfaceBreakdown";
import type { RouteSegment } from "./routeSegment";
import type { ScoreBreakdown } from "./scoreBreakdown";
import type { Waypoint } from "./waypoint";

export interface GeneratedRoute {
  id: string;
  name: string;
  description: string;
  distanceMiles: number;
  estimatedDurationMinutes: number;
  elevationGainFt: number;
  elevationLossFt: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  overallScore: number;
  scoreBreakdown: ScoreBreakdown;
  waypoints: Waypoint[];
  segments: RouteSegment[];
  surfaceBreakdown: GeneratedRouteSurfaceBreakdown;
  warnings: string[];
  highlights: string[];
}