# lib/api-zod/src/generated/types/routeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteRequestTimeOfDay } from "./routeRequestTimeOfDay";
import type { TrainingGoal } from "./trainingGoal";

export interface RouteRequest {
  trainingGoal: TrainingGoal;
  /**
   * @minimum 0.5
   * @maximum 30
   */
  distanceMiles: number;
  startLat: number;
  startLng: number;
  temperatureF?: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  humidity?: number;
  windSpeedMph?: number;
  /**
   * @minimum 0
   * @maximum 11
   */
  uvIndex?: number;
  timeOfDay?: RouteRequestTimeOfDay;
  preferShade?: boolean;
  avoidTraffic?: boolean;
  preferTrails?: boolean;
}