# lib/api-zod/src/generated/types/routeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedRoute } from "./generatedRoute";
import type { WeatherSummary } from "./weatherSummary";

export interface RouteResponse {
  routes: GeneratedRoute[];
  weatherSummary: WeatherSummary;
  generatedAt: Date;
}