# lib/api-zod/src/generated/types/routeSegment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteSegmentTrafficLevel } from "./routeSegmentTrafficLevel";
import type { SurfaceType } from "./surfaceType";

export interface RouteSegment {
  startIndex: number;
  endIndex: number;
  distanceMiles: number;
  elevationGainFt: number;
  grade: number;
  surfaceType: SurfaceType;
  shadePercentage: number;
  trafficLevel: RouteSegmentTrafficLevel;
  /**
   * @minimum 0
   * @maximum 10
   */
  safetyScore: number;
  lightingAvailable: boolean;
}