# lib/api-zod/src/generated/types/runRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TrainingGoal } from "./trainingGoal";

export interface RunRecord {
  id: number;
  routeId: string;
  trainingGoal: TrainingGoal;
  distanceMiles: number;
  durationMinutes: number;
  avgPaceMinsPerMile: number;
  elevationGainFt: number;
  temperatureF: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  perceivedEffort: number;
  notes?: string;
  completedAt: Date;
}