# lib/api-zod/src/generated/types/scoreBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBreakdown {
  terrainMatch: number;
  safetyScore: number;
  environmentalFit: number;
  trainingEffectiveness: number;
  shadeScore: number;
  trafficScore: number;
}