# lib/api-zod/src/generated/types/surfaceType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunRoute - Smart Running Route Planner API
 * OpenAPI spec version: 0.1.0
 */

export type SurfaceType = (typeof SurfaceType)[keyof typeof SurfaceType];

export const SurfaceType = {
  pavement: "pavement",
  gravel: "gravel",
  trail: "trail",
  mixed: "mixed",
} as const;